pyproject.toml
[tool.poetry]
name = "cf-bot"
version = "0.1.0"
description = "Telegram notifications bot for codeforces rounds"
authors = ["Sergey <sergejananev78@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-codeforces="^0.2.7"
python-telegram-bot = { version = "^22.0", extras = ["job-queue"] }

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
